Implement very simple Flask server
