+++ apps/hash-agents/agents/__main__.py
@@ -1,20 +1,6 @@
 import json
-import runpy
 import sys
-import importlib
-from . import setup
-
-
-def call_agent(agent: str, **kwargs) -> dict:
-    module = f"agents.{agent}"
-
-    io_types = importlib.import_module(f"{module}.io_types")
-
-    out = runpy.run_module(module, run_name='HASH', init_globals={
-        'IN': io_types.input_from_dict(dict(kwargs)),
-    }).get('OUT')
-
-    return io_types.output_to_dict(out)
+from . import setup, call_agent
 
 
 def main():
