+++ apps/hash-agents/requirements.txt
@@ -1,3 +1,4 @@
+flask==2.2.3
 langchain==0.0.137
 openai==0.27.4
 python-dotenv==1.0.0
